remove api folder
