+++ src/iceberg/CMakeLists.txt
@@ -15,7 +15,21 @@
 # specific language governing permissions and limitations
 # under the License.
 
-install(DIRECTORY "${CMAKE_CURRENT_SOURCE_DIR}/iceberg"
-        DESTINATION "${ICEBERG_INSTALL_INCLUDEDIR}"
-        FILES_MATCHING
-        PATTERN "*.h")
+set(ICEBERG_SOURCES demo_table.cc)
+set(ICEBERG_INCLUDES "${CMAKE_SOURCE_DIR}/src")
+
+add_iceberg_lib(iceberg
+                SOURCES
+                ${ICEBERG_SOURCES}
+                OUTPUTS
+                ICEBERG_LIBRARIES
+                PRIVATE_INCLUDES
+                ${ICEBERG_INCLUDES})
+
+iceberg_set_export_definitions(iceberg_static "${ICEBERG_LIBRARIES}")
+iceberg_install_all_headers(iceberg)
+
+add_subdirectory(arrow)
+add_subdirectory(puffin)
+
+iceberg_install_cmake_package(Iceberg iceberg_targets)
