+++ src/iceberg/puffin/demo_puffin.h
@@ -17,12 +17,17 @@
  * under the License.
  */
 
-#include "demo_puffin.h"
+#pragma once
 
-namespace iceberg {
+#include "iceberg/puffin.h"
 
-std::string_view DemoPuffin::print() const { return "DemoPuffin"; }
+namespace iceberg::puffin {
 
-std::unique_ptr<Puffin> Puffin::create() { return std::make_unique<DemoPuffin>(); }
+class ICEBERG_EXPORT DemoPuffin : public Puffin {
+ public:
+  DemoPuffin() = default;
+  ~DemoPuffin() override = default;
+  std::string print() const override;
+};
 
-}  // namespace iceberg
+}  // namespace iceberg::puffin
